Rename helpers to constants
